[refactor] Move the swarm's reporting surface into its own module
The snapshot builders swarm.js assembles into its status object are pure
formatters over the live Hyperswarm, so they lift out whole. The module
takes accessors rather than the handle, because initSwarm and destroySwarm
reassign it, and imports nothing from bare-* so it loads under node and is
unit tested there. DHT_VERSION stays behind in swarm.js with the bare-fs
read it needs and is passed in. swarm.js drops 142 lines; destroySwarm is
untouched, because this step moves reporting, not ownership.
